build: remove vendor prefix lint rule
We have a lint rule that tries to flag places where we were missing vendor prefixes. The problem is that it depends on very old version of `autoprefixer` and updating to a newer one will be non-trivial.

These changes remove the rule, because we didn't seem to have been getting much benefit out of it anyway, considering that browsers haven't been introducing as many vendor-prefixed APIs in the past few years.